final model fixing

+++ cg_lims/models/arnold/prep/microbial_prep/post_pcr_bead_purification.py
@@ -20,7 +20,7 @@ class PostPCRBeadPurificationProcessUDFS(BaseModel):
 class PostPCRBeadPurificationArtifactUDF(BaseModel):
     finished_library_concentration: float = Field(..., alias="Concentration")
     finished_library_concentration_nm: float = Field(..., alias="Concentration (nM)")
-    finished_library_size: Optional[float] = Field(None, alias="Size (bp)")
+    finished_library_size: Optional[int] = Field(None, alias="Size (bp)")
     finished_library_average_size: float = Field(..., alias="Average Size (bp)")
